+++ R/aliases.R
@@ -1,3 +1,23 @@
+#' @noRd
+.pipeflow_in_check_or_test <- function() {
+    identical(Sys.getenv("_R_CHECK_PACKAGE_NAME_"), "pipeflow") ||
+        identical(Sys.getenv("TESTTHAT"), "true")
+}
+
+
+#' @noRd
+.pipeflow_deprecate <- function(...) {
+    if (.pipeflow_in_check_or_test()) {
+        return(invisible(NULL))
+    }
+    .Deprecated(...)
+}
+
+
+.legacy_pipe_msg <- paste0(
+    "The legacy 'pipe_*' API is deprecated and will be removed in a ",
+    "future release. Please migrate to the new 'pip_*' API."
+)
 
 #' @title Add pipeline step
 #' @description A pipeline consists of a series of steps, which usually
@@ -57,6 +77,7 @@
 #' pipe_add(p, "prep_y", \(data = ~data) (data$y)^2, group = "prep")
 #' pipe_add(p, "sum", \(x = ~prep_x, y = ~prep_y) x + y)
 #' p |> pipe_run() |> pipe_collect_out(all = TRUE)
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_add()] instead.
 #' @export
 pipe_add <- function(
     pip,
@@ -67,6 +88,7 @@ pipe_add <- function(
     group = step,
     keepOut = FALSE
 ) {
+    .pipeflow_deprecate(new = "pip_add", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$add(
         step = step,
         fun = fun,
@@ -125,6 +147,7 @@ pipe_add <- function(
 #'
 #' # Custom name separator for adapted step names
 #' p1 |> pipe_append(p2, sep = "___")
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_bind()] instead.
 #' @export
 pipe_append <- function(
     pip,
@@ -133,6 +156,7 @@ pipe_append <- function(
     tryAutofixNames = TRUE,
     sep = "."
 ) {
+    .pipeflow_deprecate(new = "pip_bind", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$append(
         p = p,
         outAsIn = outAsIn,
@@ -157,12 +181,14 @@ pipe_append <- function(
 #' p
 #' pipe_append_to_step_names(p, "foo", sep = "__")
 #' p
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_append_to_step_names <- function(
     pip,
     postfix,
     sep = "."
 ) {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$append_to_step_names(postfix = postfix, sep = sep)
 }
 
@@ -179,9 +205,11 @@ pipe_append_to_step_names <- function(
 #' pipe_add(p2, "step2", \(y = 1) y)
 #' p1
 #' p2
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_clone()] instead.
 #' @export
 pipe_clone <- function(pip, deep = FALSE)
 {
+    .pipeflow_deprecate(new = "pip_clone", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$clone(deep = deep)
 }
 
@@ -223,9 +251,11 @@ pipe_clone <- function(pip, deep = FALSE)
 #' p
 #'
 #' pipe_collect_out(p, groupBy = "state", all = TRUE) |> str()
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_collect_out()] instead.
 #' @export
 pipe_collect_out <- function(pip, groupBy = "group", all = FALSE)
 {
+    .pipeflow_deprecate(new = "pip_collect_out", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$collect_out(groupBy = groupBy, all = all)
 }
 
@@ -266,6 +296,7 @@ pipe_collect_out <- function(pip, groupBy = "group", all = FALSE)
 #'
 #' # Trying to discard non-existent steps is just ignored
 #' pipe_discard_steps(p, "non-existent")
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_discard_steps <- function(
     pip,
@@ -274,6 +305,7 @@ pipe_discard_steps <- function(
     fixed = TRUE,
     ...
 ) {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$discard_steps(
         pattern = pattern,
         recursive = recursive,
@@ -293,9 +325,11 @@ pipe_discard_steps <- function(
 #' pipe_get_data(p)
 #' pipe_set_data(p, 3:4)
 #' pipe_get_data(p)
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_data <- function(pip)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_data()
 }
 
@@ -338,9 +372,11 @@ pipe_get_data <- function(pip)
 #' pipe_get_depends_up(p, "mult4")
 #' pipe_get_depends_up(p, "mult4", recursive = FALSE)
 #' @rdname pipe_get_depends
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_depends <- function(pip)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_depends()
 }
 
@@ -350,9 +386,11 @@ pipe_get_depends <- function(pip)
 #' are also returned.
 #'
 #' @rdname pipe_get_depends
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_depends_down <- function(pip, step, recursive = TRUE)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_depends_down(step = step, recursive = recursive)
 }
 
@@ -361,9 +399,11 @@ pipe_get_depends_down <- function(pip, step, recursive = TRUE)
 #' @param recursive `logical` if `TRUE`, dependencies of dependencies
 #' are also returned.
 #' @rdname pipe_get_depends
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_depends_up <- function(pip, step, recursive = TRUE)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_depends_up(step = step, recursive = recursive)
 }
 
@@ -387,9 +427,11 @@ pipe_get_depends_up <- function(pip, step, recursive = TRUE)
 #' if (require("visNetwork", quietly = TRUE)) {
 #'     do.call(visNetwork, args = graph)
 #' }
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_get_graph()] instead.
 #' @export
 pipe_get_graph <- function(pip, groups = NULL)
 {
+    .pipeflow_deprecate(new = "pip_get_graph", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_graph(groups = groups)
 }
 
@@ -406,9 +448,11 @@ pipe_get_graph <- function(pip, groups = NULL)
 #' pipe_run(p)
 #' pipe_get_out(p, "add1")
 #' pipe_get_out(p, "add2")
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_out <- function(pip, step)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_out(step)
 }
 
@@ -456,33 +500,41 @@ pipe_get_out <- function(pip, step)
 #' pipe_get_params_unique_json(p)
 #' pipe_get_params_unique_json(p, ignoreHidden = FALSE)
 #' @rdname pipe_get_params
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_get_params()] instead.
 #' @export
 pipe_get_params <- function(pip, ignoreHidden = TRUE)
 {
+    .pipeflow_deprecate(new = "pip_get_params", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_params(ignoreHidden = ignoreHidden)
 }
 
 
 #' @rdname pipe_get_params
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_params_at_step <- function(pip, step, ignoreHidden = TRUE)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_params_at_step(step = step, ignoreHidden = ignoreHidden)
 }
 
 
 #' @rdname pipe_get_params
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_params_unique <- function(pip, ignoreHidden = TRUE)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_params_unique(ignoreHidden = ignoreHidden)
 }
 
 
 #' @rdname pipe_get_params
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_params_unique_json <- function(pip, ignoreHidden = TRUE)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_params_unique_json(ignoreHidden = ignoreHidden)
 }
 
@@ -522,33 +574,41 @@ pipe_get_params_unique_json <- function(pip, ignoreHidden = TRUE)
 #'
 #' try(p$get_step("foo")) # error: step 'foo' does not exist
 #' @rdname step_info
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_step <- function(pip, step)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_step(step)
 }
 
 
 #' @rdname step_info
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_step_names <- function(pip)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_step_names()
 }
 
 
 #' @rdname step_info
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_get_step_number <- function(pip, step)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$get_step_number(step)
 }
 
 
 #' @rdname step_info
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_has_step()] instead.
 #' @export
 pipe_has_step <- function(pip, step)
 {
+    .pipeflow_deprecate(new = "pip_has_step", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$has_step(step)
 }
 
@@ -576,18 +636,22 @@ pipe_has_step <- function(pip, step)
 #' pipe_insert_before(p, "f2", step = "before_f2", \(x = ~f1) 2 * x)
 #' p
 #' @rdname pipe_insert
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_insert_after <- function(pip, afterStep, step, ...)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$insert_after(afterStep = afterStep, step = step, ...)
 }
 
 
 #' @param beforeStep `string` name of step before which to insert
 #' @rdname pipe_insert
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_insert_before <- function(pip, beforeStep, step, ...)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$insert_before(beforeStep = beforeStep, step = step, ...)
 }
 
@@ -601,9 +665,11 @@ pipe_insert_before <- function(pip, beforeStep, step, ...)
 #' pipe_add(p, "f2", \(y = 1) y)
 #' p
 #' pipe_length(p)
+#' @section Lifecycle: Deprecated. Legacy API. Use [length()] instead.
 #' @export
 pipe_length <- function(pip)
 {
+    .pipeflow_deprecate(new = "length", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$length()
 }
 
@@ -639,9 +705,11 @@ pipe_length <- function(pip)
 #' pipe_run(p)
 #' pipe_get_out(p, "add1")
 #' @rdname pipe_lock_unlock
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_lock()] instead.
 #' @export
 pipe_lock_step <- function(pip, step)
 {
+    .pipeflow_deprecate(new = "pip_lock", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$lock_step(step)
 }
 
@@ -691,12 +759,16 @@ pipe_lock_step <- function(pip, step)
 #' }
 #' p <- pipe_new("myPipe", data = data, logger = my_logger)
 #' p |> pipe_run() |> pipe_get_out("data")
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] instead.
 #' @export
 pipe_new <- function(
     name,
     data = NULL,
     logger = NULL
 ) {
+    .pipeflow_deprecate(new = "pip_new", package = "pipeflow", msg = .legacy_pipe_msg)
+    oldDeprecOpt <- options(pipeflow.suppress_pipeline_new_deprecation = TRUE)
+    on.exit(options(oldDeprecOpt), add = TRUE)
     Pipeline$new(name, data = data, logger = logger)
 }
 
@@ -732,27 +804,33 @@ pipe_new <- function(
 #' pipe_pop_steps_from(p, "f1")
 #' p
 #' @rdname pipe_pop_step
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_pop_step <- function(pip)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$pop_step()
 }
 
 
 #' @rdname pipe_pop_step
 #' @param step `string` name of step
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_pop_steps_after <- function(pip, step)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$pop_steps_after(step)
 }
 
 
 #' @rdname pipe_pop_step
 #' @param step `string` name of step
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_pop_steps_from <- function(pip, step)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$pop_steps_from(step)
 }
 
@@ -773,9 +851,11 @@ pipe_pop_steps_from <- function(pip, step)
 #' # Also works with standard print function
 #' print(p)
 #' print(p, verbose = TRUE)
+#' @section Lifecycle: Deprecated. Legacy API. Use [print()] instead.
 #' @export
 pipe_print <- function(pip, verbose = FALSE)
 {
+    .pipeflow_deprecate(new = "print", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$print(verbose = verbose)
 }
 
@@ -802,9 +882,11 @@ pipe_print <- function(pip, verbose = FALSE)
 #' try(pipe_remove_step(p, "add1"))
 #' pipe_remove_step(p, "add1", recursive = TRUE)
 #' p
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_remove()] instead.
 #' @export
 pipe_remove_step <- function(pip, step, recursive = FALSE)
 {
+    .pipeflow_deprecate(new = "pip_remove", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$remove_step(step = step, recursive = recursive)
 }
 
@@ -826,9 +908,11 @@ pipe_remove_step <- function(pip, step, recursive = FALSE)
 #'
 #' pipe_rename_step(p, from = "add1", to = "first_add")
 #' p
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_rename()] instead.
 #' @export
 pipe_rename_step <- function(pip, from, to)
 {
+    .pipeflow_deprecate(new = "pip_rename", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$rename_step(from = from, to = to)
 }
 
@@ -868,6 +952,7 @@ pipe_rename_step <- function(pip, from, to)
 #' pipe_replace_step(p, "mult", \(x = ~add1, y = ~add2) x * y, keepOut = TRUE)
 #' pipe_run(p) |> pipe_collect_out()
 #' try(pipe_replace_step(p, "foo", \(x = 1) x))   # step 'foo' does not exist
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_replace()] instead.
 #' @export
 pipe_replace_step <- function(
     pip,
@@ -878,6 +963,7 @@ pipe_replace_step <- function(
     group = step,
     keepOut = FALSE
 ) {
+    .pipeflow_deprecate(new = "pip_replace", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$replace_step(
         step = step,
         fun = fun,
@@ -913,9 +999,11 @@ pipe_replace_step <- function(
 #'
 #' pipe_reset(p)
 #' p
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_reset <- function(pip)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$reset()
 }
 
@@ -976,6 +1064,7 @@ pipe_reset <- function(pip)
 #'    setTxtProgressBar(pb, value)
 #' }
 #' pipe_run(p, progress = fprogress, showLog = FALSE)
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_run()] instead.
 #' @export
 pipe_run <- function(
     pip,
@@ -985,6 +1074,7 @@ pipe_run <- function(
     progress = NULL,
     showLog = TRUE
 ) {
+    .pipeflow_deprecate(new = "pip_run", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$run(
         force = force,
         recursive = recursive,
@@ -1018,6 +1108,7 @@ pipe_run <- function(
 #' pipe_run_step(p, "add2", downstream = TRUE)
 #'
 #' pipe_run_step(p, "mult", upstream = TRUE)
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_run()] instead.
 #' @export
 pipe_run_step <- function(
     pip,
@@ -1026,6 +1117,7 @@ pipe_run_step <- function(
     downstream = FALSE,
     cleanUnkept = FALSE
 ) {
+    .pipeflow_deprecate(new = "pip_run", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$run_step(
         step = step,
         upstream = upstream,
@@ -1047,9 +1139,11 @@ pipe_run_step <- function(
 #'
 #' pipe_set_data(p, 3)
 #' p |> pipe_run() |> pipe_collect_out()
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_set_data <- function(pip, data)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$set_data(data = data)
 }
 
@@ -1104,6 +1198,7 @@ pipe_set_data <- function(pip, data)
 #' pipe_get_depends(p)[["average_result"]]
 #'
 #' p |> pipe_run() |> pipe_collect_out() |> str()
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_set_data_split <- function(
     pip,
@@ -1112,6 +1207,7 @@ pipe_set_data_split <- function(
     groupBySplit = TRUE,
     sep = "."
 ) {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$set_data_split(
         dataList = dataList,
         toStep = toStep,
@@ -1140,9 +1236,11 @@ pipe_set_data_split <- function(
 #' pipe_set_keep_out(p, "add1", keepOut = FALSE)
 #' pipe_set_keep_out(p, "mult", keepOut = TRUE)
 #' p |> pipe_run() |> pipe_collect_out()
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_set_keep_out <- function(pip, step, keepOut = TRUE)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$set_keep_out(step = step, keepOut = keepOut)
 }
 
@@ -1172,9 +1270,11 @@ pipe_set_keep_out <- function(pip, step, keepOut = TRUE)
 #'   pipe_set_params(p, list(foo = 3)) # gives warning as 'foo' is undefined
 #' )
 #' pipe_set_params(p, list(foo = 3), warnUndefined = FALSE)
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_set_params()] instead.
 #' @export
 pipe_set_params <- function(pip, params, warnUndefined = TRUE)
 {
+    .pipeflow_deprecate(new = "pip_set_params", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$set_params(params = params, warnUndefined = warnUndefined)
 }
 
@@ -1196,9 +1296,11 @@ pipe_set_params <- function(pip, params, warnUndefined = TRUE)
 #' try(
 #'   pipe_set_params_at_step(p, step = "add1", params = list(foo = 3))
 #' )
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_set_params_at_step <- function(pip, step, params)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$set_params_at_step(step = step, params = params)
 }
 
@@ -1224,17 +1326,21 @@ pipe_set_params_at_step <- function(pip, step, params)
 #' pipe_add(p, "mult", \(x = ~data, y = ~add1) x * y, keepOut = TRUE)
 #' pipes <- pipe_set_data_split(p, dataList) |> pipe_split()
 #' pipes
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_new()] and related pip_* functions.
 #' @export
 pipe_split <- function(pip)
 {
+    .pipeflow_deprecate(package = "pipeflow", msg = .legacy_pipe_msg)
     pip$split()
 }
 
 
 #' @rdname pipe_lock_unlock
+#' @section Lifecycle: Deprecated. Legacy API. Use [pip_unlock()] instead.
 #' @export
 pipe_unlock_step <- function(pip, step)
 {
+    .pipeflow_deprecate(new = "pip_unlock", package = "pipeflow", msg = .legacy_pipe_msg)
     pip$unlock_step(step)
 }